Começamos a trabalhar com Jest. .. . . . . . . .

+++ src/tasks/tasks.service.ts
@@ -3,6 +3,7 @@ import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { PaginationDto } from 'src/common/dto/pagination.dto';
+import { PayloadTokenDto } from 'src/auth/dto/payload-token.dto';
 
 @Injectable()
 export class TasksService {
@@ -31,14 +32,14 @@ export class TasksService {
         throw new HttpException('Tarefa não encontrada', HttpStatus.NOT_FOUND);
     }
 
-    async create(createTaskDto: CreateTaskDto) {
+    async create(createTaskDto: CreateTaskDto, tokenPayLoad: PayloadTokenDto) {
         try {
             const newTask = await this.prisma.task.create({
             data: {
                     name: createTaskDto.name,
                     description: createTaskDto.description, 
                     completed: false,
-                    userId: createTaskDto.userId,
+                    userId: tokenPayLoad.sub,
                 }, select: {
                     id: true,
                     name: true,
@@ -53,13 +54,17 @@ export class TasksService {
         }        
     }
 
-    async update(id: string, updateTaskDto: UpdateTaskDto) {
+    async update(id: string, updateTaskDto: UpdateTaskDto, tokenPayLoad: PayloadTokenDto) {
         const findTask = await this.prisma.task.findFirst({
             where: { id: Number(id) }
         });
 
         if (!findTask?.name) {
             throw new HttpException('Tarefa não encontrada', HttpStatus.NOT_FOUND);
+        }
+
+        if(findTask.userId !== tokenPayLoad.sub) {
+            throw new HttpException('Você não tem permissão para atualizar esta tarefa', HttpStatus.FORBIDDEN);
         }
 
         const updatedTask = await this.prisma.task.update({
@@ -70,7 +75,7 @@ export class TasksService {
         return updatedTask;
     }
 
-    async remove(id: string) {
+    async remove(id: string, tokenPayLoad: PayloadTokenDto) {
         try {
             const findTask = await this.prisma.task.findFirst({
                 where: { id: Number(id) }
@@ -80,13 +85,17 @@ export class TasksService {
                 throw new HttpException('Tarefa não encontrada', HttpStatus.NOT_FOUND);
             }
 
+            if(findTask.userId !== tokenPayLoad.sub) {
+                throw new HttpException('Você não tem permissão para deletar esta tarefa', HttpStatus.FORBIDDEN);
+            }
+
             await this.prisma.task.delete({
                 where: { id: Number(id) }
             });
 
             return { message: 'Tarefa deletada com sucesso!' };    
         } catch (error) {
-            throw new HttpException('Erro ao deletar a tarefa. Mensagem: ' + error, HttpStatus.INTERNAL_SERVER_ERROR);    
+            throw new HttpException(error, error.status);    
         }
         
     }
